Order item

+++ src/main/java/com/devshop/devshop/repository/ProductRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
     List<Product> findByCategoryId(Long id);
+
 }
